Performance measurement. Creation of first gnuplot script

+++ Java/test/org/hpms/mw/distcrud/samples/Settings.java
@@ -2,7 +2,7 @@ package org.hpms.mw.distcrud.samples;
 
 public interface Settings {
 
-   String MC_GROUP  = "224.0.0.3";
-   String MC_INTRFC = "eth1";
-   int    MC_PORT   = 2416;
+   String MC_GROUP    = "224.0.0.3";
+   String MC_INTRFC   = "eth1";
+   short  MC_PORT     = 2416;
 }
